requirements.txt
python-telegram-bot>=20.0
fastapi-poe>=0.0.40
requests
google-generativeai>=0.6.0
pillow
asyncio 